Bring out install() to import to two places
